Last minute changes

+++ common/src/inventory/item.rs
@@ -3,6 +3,7 @@ use specs::{Component, VecStorage};
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub enum Armor {
+    //TODO: Don't make armor be a bodypart. Wearing enemy's head is funny but also creepy thing to do.
     Helmet(actor::Head),
     Shoulders(actor::Shoulder),
     Chestplate(actor::Chest),
@@ -27,14 +28,14 @@ pub enum Rarity {
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub enum Item {
     Weapon {
-        damage: i16,
-        strength: i16,
+        damage: i32,
+        strength: i32,
         rarity: Rarity,
         variant: actor::Weapon,
     },
     Armor {
-        defense: i16,
-        health_bonus: i16,
+        defense: i32,
+        health_bonus: i32,
         rarity: Rarity,
         variant: Armor,
     },
